Extract joi-extensions to standalone files
